requirements.txt
# local package
# -e .

# external requirements
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv>=0.5.1
